requirements.txt
flask-socketio
Flask-APScheduler
gevent-websocket
numpy
matplotlib
seaborn
